requirements.txt
streamlit_option_menu
streamlit_nested_layout
scholarly
cachetools
sniffio
lxml
